[1861/1867] Better help text for reduce token step
A player will need to remove tokens from the map during the merger
process if the merging minors have more than two tokens on the map, or
if there are multiple tokens in the same hex (or both of these). The
help text shown only covered the first of these so could be misleading.

This changes to show a message that more accurately describes why tokens
need to be removed.

Fixes part of tobymao/18xx#8589.