[1861/1867] Check the player isn't removing too many tokens in a merger
In the reduce token step, if there were three or more minors merging and
there were two tokens in the same hex, it was possible for the player to
remove all but one token from the map. This is incorrect, the new public
company should be formed with two tokens in different hexes. The only
case where the new public company can have just one token is if all the
merging minors had tokens in the same hex.

This commit adds a check that the player isn't trying to remove too many
tokens and throws an error if they are.

Fixes the last final part of tobymao/18xx#8589.