Refactor: remove_reservation! instead of reservations.delete
This is in preparation for changing the type of object that will be held
in the City.reservations and Tile.reservations arrays.

Calling reservations.delete is also potentially buggy: if there are
multiple reservations for a city then calling delete can change the
slots that other reservations point too. City.remove_reservation! is
better as it swaps in a nil value instead of deleting the element from
the reservations array.