Create a Reservation tile part for reserved token slots
The `reservations` attribute of City and Tile objects is currently an
array of Operator objects, representing the companies, minors and
corporations that have a reserved token space.

This commit creates a Part::Reservation class to use with these
`reservations` lists. This commit is the minimal amount of change to
switch over to using these Reservation objects. Later commits will add
more attributes and functionality to the Reservation class.