[1858] Check for consent in the private closure round
In the private closure round, if a private is exchanged for a treasury
share then consent is needed from the corporation's owner, if this is
not the same player as the private's owner. Consent is not needed when
exchanging for a market share.

To implement this, a method in Game::Base is created called
`consenter_for_choice`, which allows a game to check whether consent is
needed for any Choose actions. The default implemention just returns
nil, which means that no consent is required.