[1858] Use correct token when loading game
The cost of a token depends on the number of provincial borders crossed,
and is calculated in Step::Token.token_cost_override!, but this method
is only called from the game view. This meant that the default cost of
Pt20 was being used for token placements when the game was loaded,
causing games to break.

To fix this, we need to use the cost stored in the place_token action.