[1858] Better options for selecting variants
The two quick start variants are now mutually exclusive, you either
select the set A option or the set B option. This should be clearer than
having to select both options to get the set B quick start variant.

This makes it impossible to select the quick start variants for a
two-player game.