Improved game end message when bank breaks in stock round
For a game with a game end timing of :current_or, if the bank breaks
during a stock round you finish the stock round and play one operating
round. But the game end message is a bit confusing during the stock
round, it says something like:

> Stock Round 7 - Sell then Buy Shares - Bank Broken : Game Ends at conclusion of this OR (7.1)

This changes the message to say 'the next OR' if we are not already in
an operating round.