Improved message in emergency money raising
For games where player shares cannot be sold in emergency money raising,
the message only shows how much the president can contribute to afford
the cheapest available train, even if more can be spent to buy a more
expensive train.

This adds a message showing if more can be contributed to buy a
different train from the depot.