[1822CA] P26-P27 Grain Trains, grain elevators, and ports
Grain trains can attach to regular trains in order to earn revenue at grain
elevators without counting the town as a stop. All standard trains except
E-trains can earn bonuses at grain elevators, and an extra bonus if the route
includes a grain elevator and a port.

Any train except an E-train can attach a grain train. This is done in the Route
step. If a Major owns both a pullman and a grain train, the pullman is
attached (or skipped) first.

Details on the revenue bonus rules can be found in the
[rulebook](https://boardgamegeek.com/filepage/238950/1822ca-rules):

* the table in rule 5.11.27 on page 21 covers most of it
* 5.11.25 shows the port limit on page 21

Very small, semi-related changes also included here:

* reorder pullman's distance to fix it counting stops; this was broken when
  enough towns were added since it was originally copied from 1822 pullmans,
  which have no limit to the number of towns
* after the train object is created, rename the "5P" train to "5"; the different
  id/name is useful in setup, but once a Major owns the train it is just like
  any other 5-train (and IMO looks better when renamed from attaching the
  pullman or grain train)

Core changes:

* extract new method `stop_type()` in `Game::Base`, so that grain elevator towns
  could be treated as the new type `grain_elevator` instead of `town` for grain
  trains
* in `Engine::Train`, added `attr_accessor :no_local` so it could be toggled
  when attaching a grain train; this was necessary for the LP-G train to select
  more than 2 stops