[1858] Require explicit approval/denial of exchange requests in PCR
In the private closure round, exchanging a private railway company for a
treasury share is only allowed if the public company's president
approves the exchange request.

In addition, if there a public company has shares available in both its
treasury and the market, then a private can only be exchanged for a
market share if the public company's director has first refused an offer
to exchange for a treasury share.

This is all a bit too complicated to rely on players correctly using the
confirm dialog, so instead an exchange request is passed on to the
public company's president. They are given 'approve' and 'deny' buttons
to respsond to the request. If they approve then the swap happens. If
they do not approve then control passes back to the private railway's
owner, and they now get the option to exchange for any of the public
company's market shares.