[18Ardennes] Guillaume-Luxembourg can be sold and bought in SRs
The Guillaume-Luxembourg can be sold back to the bank for 100F. And, if
it is owned by the bank, another player can buy it for 100F.

Also hide the other minor railways (M1–M15) in the stock round as you
can't do anything with them (at the moment, anyway). And stop major
companies from being parred, they can't be started directly.