[18Ardennes] Add special handling for the Ruhr green tile
The Ruhr hex starts of with a preprinted yellow tile containing a city
and a town. In green this upgrades to a single city with two slots.
This normal code rejects this tile as a valid upgrade because of the
disappearing town, so add special cases to allow this tile to be placed.