Add `exta_revenue` attribute to Action::RunRoutes
18Ardennes has additional revenue that is earned on top of the normal
train revenue. This revenue comes tokens collected during the game, the
value of each is determined by the corporation's routes, so the game
graph needs to be consulted. This means that the value shouldn't be
calculated during the dividend step, as this would force recomputing the
game graph many times whilst a game is loaded.

This commit adds a new attribute, `extra_revenue`, to Action::RunRoutes.
This is set by a call from `View::Game::RouteSelector` to
`Game::Base.extra_revenue`. This value is stored in the game JSON, so
can be read when loading a game without needed to compute the game
graph.