[18Ardennes] Major auction round
A simultaneous auction for the rights to start public companies in the
following stock rounds. This is implemented by auctioning companies of
type :concession, one for each major corporation.

When a bid is made, the player must nominate one of their minors to be
folded into the major. This is done by giving them several bid buttons,
one for each eligible minor.