[1858] Make sure token from private goes to correct city
Catch a corner case: if any of the Madrid private railway companies are being
acquired by a public company and one of the other Madrid slots is empty and
unreserved (after another Madrid private closed without their slot being taken),
then the player can select either the slot for the private being acquired or the
empty slot. Only the first of these is a legal choice.

Fixes tobymao#9858.