[18Ardennes] Only allow pool share exchange after treasury share refused
If a public company has shares both in its treasury and in the bank
pool, then another player can only exchange a minor company for the pool
share if they have first attempted the exchange for a treasury share,
and this request was refused by the public company's president.

To do this track when a request for a treasury share exchange has been
refused, and only offer the option of exchanging for a pool share if a
request refusal has been received.

The game rules aren't clear on what should happen if there are only
pool shares available, and none in the treasury. This assumes that you
can exchange without needing the president refuse an impossible
exchange.