[18Ardennes] Change minor tokens to show associated public companies
In the first round of starting public companies, these can only be
started using minro companies that have tokens in certain cities.

To make the association between cities and public comanies visible to
players, this commit changes minor companies' tokens when they are
placed in cities, to give them the colours of the associated public
companies.

- A token in Rotterdam will be yellow, as this one could only be used to
  start the NL public company.
- A token in Lille-Roubaix will be half green and half brown, as this is
  associated with both the BE and N public companies.
- A token in Aachen will be yellow, green and dark blue, as all of NL,
  BE and P can be started from there.