[18CH] Mountain railway tiles
These tiles can be laid on big (120sfr cost) mountains and are an
off-board area with 30sfr revenue. The corporation that builds these get
a +40 token, and each 40sfr extra at each run routes step as long as
they are running a train. This additional income is stored in the
`extra_revenue` attribute (similar to 18Ardennes fort bonuses) as it is
not tied directly to a route.