[1858] Prevent creation of extra token slot in cities
tobymao#10443 was an incomplete fix for tobymao#10442. The fix only
covered the case where the public company (corporation) took the the
private railway (private) token slot when it was first formed and placed
its home token. It did not cover the case where a private railway was
later exchanged for a public company share and the public company then
took the token slot. This was seen in ollybh#142.

This now makes sure that the private railway token reservation is
deleted in either case.