[18CH] Fix error when no mountain railway tiles left
If someone was trying to place a tile in a big mountain hex after all
the mountain railway tiles had been used, an error was occurring which
prevented any tiles from being played in that hex. This was because nil
was being returned by `tiles.find` and appended to the
`all_potential_tiles` array.