[18CH] Special rules for privates with Gotthard as home hex
In 1858 if a plain track hex is one of a private railway’s home hexes,
then they can trace a route out of the hex using any track in that hex.

The Gotthard hex (H11) in 18CH is different. It is a pre-printed grey
hex with two pieces of track, one broad gauge and one metre (narrow)
gauge. Two private railways have Gotthard as on of their home hexes. GB
is only allowed to use the broad gauge track to trace routes, and FOB is
only allowed to use the metre gauge track.