[18CH] Extra checks for exchanging FOB/GB privates
There are extra checks to be made before one of the FOB or GB privates
can be exchanged for a public company’s share, if the public company has
a connection to the private’s home in H11 (Gotthard). The FOB can only
be exchanged if the public company has a connection to the narrow gauge
track in H11, the GB if there is a connection to the broad gauge track.