[18Ardennes] The GL minor can be sold when over cert limit
The game was only allowing a player to sell public company shares when
they were over the certificate limit. If they own the GL minor this
should also be available to be sold. This wasn't possible as it is
implemented as a private company, and the inherited implementation
wasn't including `sell_company` as an action.

Fixes ollybh#136.