[18Ardennes] Bankrupt a player without cert slot to start a public company
It's theoretically possible to get in the position where you have the
obligation to start a public company, but you do not have the free
certificate slot to be able to do this, and have nothing that can be
sold to free up a slot. If that happens then you go bankrupt.