[18Ardennes] Don't allow excess sales before starting public company
When a player has won the right to start a public company, they must do
this on their first turn in the next stock round (or first turns if they
won multiple auctions). They are allowed to sell shares (or the GL
minor) before starting the public company, but only if the money raised
goes towards starting the new company.

This commit adds a check that the par price chosen could not have been
used if less items had been sold. If this would have been possible then
an error is thrown, making the player either choose a higher par price,
or undo some (or all) of the sales.