[1858Switzerland] Robot variant for two-players
Adds a 1858 Switzerland variant, where there is a robot player and the
SBB national railway added to the two-player game.

The robot player acquires the most expensive available private railway
company on its stock turns. It is always last in priority order.

The robot-owned private railway companies must lay track in their home
hexes in the operating turns. The players alternate choosing which tiles
to lay, with the priority holder choosing for the first company.

The SBB starts operating once the first train has been bought. It will
upgrade one of the cities where it has a token (tile chosen by the
priority holder in the first operating round of a set, and the other
player in the second), and will acquire a train in the second operating
round of a set. It does not run its trains.