[1858] Add track_type attributes to train definitions
The track_type attribute was added to Train objects for the 18ESP. Use
this for 1858's trains, instead of parsing the trains' names to work out
whether they are broad or metre (narrow) gauge trains.

Also change the check for a hex train to look at its distance attribute.
If this is a number then the train is a hex train. All non-hex trains
have hashes for their distance attribute.