[1837] Use backtracking graph in the token step
1837 has two graphs: a standard one and one that allows track to
backtrack at junctions. The latter is needed for placing station tokens,
where this type of route is allowed.

The token step was not using this graph, it was using the standard one.
This meant that not all possible token locations were being included,
and an auto-pass was being generated for the token step if the only
possible token locations involved backtracking.

Fixes tobymao#11600.