[1858] Add `private_railway?` test for company objects
In 1858 Iberia and 1858 Switzerland all company objects are part of
private railway companies. These do not behave like companies in most
other games: they do not pay owning corporations income at the start of
the operating round, instead it is added to the corporation's train
revenue.

1858 India adds locomotive works privates, which behave much more like
private companies in other games.

This adds a `private_railway?` method that always returns true. This
method is checked in various places to decide whether the company should
behave like a 1858 private railway company or a standard private
company.