[1858 India] Auto-pass when no route to mine/port tokens
Checking whether a corporation has a route to an available token
involves recalculating the game graphs, which is expensive, so `actions`
only checks if there is an available token somewhere on the board. To
avoid asking the player to choose tokens when none are really available,
insert an auto-action pass.