[1807] Bonuses for Irish ferry privates
The three Irish ferry companies give a hex bonus, but this is more
complicated than the usual case. Each of these companies can be used to
give a bonus for any of the three Ireland offboard areas, but only one
each operating turn, and if a corporation owns multiple Irish ferries
then each ones bonus must be used for a different offboard area.