[1807] Allow London to be upgraded
Because the connections between the London hex on the map and its
neighbours are severed, we need some extra steps to allow this tile to
be upgraded.

This also makes the zoomed version of the London hex start as yellow,
and updates its colour and revenue when the London tile is upgraded.