[1867] Abstract tests for types of corporation
The code in 1867 is written to handle two types of corporations: minors
and majors (aka public companies). 1807 is being implemented as an 1867
derivative and it adds a third type of corporation: systems. Systems are
more like majors than public companies.

The 1867 dividend step has tests for minors and majors. For several
methods changing the test from 'is this a major?' to 'is this not a
minor?' means that 1807 does not need to override the 1867 code.

This also adds methods to the game class to abstract the tests for minor
and major corporations.