[1867] Refactor: extract Step::Merge.connected_cities method
Refactor the code for determining corporations connected for mergers by
extracting a method that returns the cities that a corporation is
connected to.

This is being done to make the 1807 implemenation easier. It changes
this algorithm by allowing rickshaw connections between corporations
with tokens in London. Extracting this method means that only this
method will need to be overridden in the 1807 code.