[1858] Fix error in hotseat games
I wasn't able to reproduce this myself, but someone else was getting
errors when starting 1858 and 1858 India games in hotseat mode. The
error was occuring when trying to place the first bid on a private
railway company: `undefined method 'cash' for nil`.

This was occurring in the loop iterating over
`@game.programmed_actions`. It was returning an element where `player`
(the key) was nil and the `actions` (values) were an empty array.

This prevents the error by testing for the nil key.

Fixes ollybh#196.