[1867] Code change for 1807
Rewrite a check as to whether minor companies can be merged. `mergeable`
returns `mergeable_candidates` if both `@converting` and
`@merge_majors` are nil, which we know to be the case here, so change
this to use that method directly.

This change makes the 1807 implementation easier. `mergeable` might
return a hash in the 1807 code so the existing check would not work.

Fixes ollybh#198.