[Core] Improved error message on creating an empty ShareBundle
If you tried to create a ShareBundle, passing an empty array as the
`shares` parameter, then you would get an error thrown with the
description 'All shares must be from the same corporation'. That isn't
the clearest error message.

This commit changes the error description to 'A share bundle must
contain at least one share'.

It also changes any errors thrown by the ShareBundle constructor to be
GameErrors instead of RuntimeErrors.