[1867] Changed test for the end of a merger round
I'm not entirely sure why this makes a difference, but the existing code
was not always working for 1807. It was possible for
`next_entity_index!` to move the active entity back to the first
corporation in `@entities` and not have the merger round end.

Testing for having reached the end of `@entities` works.