[Route selector] Cycle connections from head and tail of chain
There are times when the player is trying to add a node (city/town/halt)
to a route and it is possible to reach this node from both ends of the
existing route chain.

The current behaviour is that only connections to the new node from the
head of the route chain will be offered.

This changes this to also offer connections from the tail of the route
chain. As the player clicks on the node, they will first be offered the
head connections, then the tail connections, and then the connection to
the new node will be removed once all possible connections have been
offered.

Fixes tobymao#3915.