[Route selector] Offer shorter connections first
When train routes are being selected and there are multiple possible
routes to the new node from the end of the route chain, these are being
offered in the order they are found when walking the graph.

There are cases where game maps can offer hundreds or even thousands of
possible connections between nodes. It is likely that the player will
like one of the more direct connections between the nodes, but these
could be a long way down the list of options, requiring a lot of
clicking to find the desired route.

This commit sorts the options by the number of paths between the nodes,
so the more direct connections are offered first.

Fixes tobymao#11774.