[1858 India] Improved message if unable to attach a mail train
In the run routes step, a mail train can be attached to a broad gauge
train to give increased revenue.

This commit shows a message if a company owns a mail train but only owns
narrow gauge trains. Before this change, there was a message asking the
player to attach the mail train to another train, but no buttons to let
them do this.

Fixes ollybh#159.