[1807] Fix game end timing
The end game check had been broken by the refactoring of the end game
code that was part of PR tobymao#12086.

Fix this, and take advantage of the new `game_end_set_final_turn!`
method that was added to remove the need for a special event to trigger
the end of the game.